requirements.txt
# apturl==0.5.2
# arabic-reshaper==3.0.0
# asgiref==3.8.1
# asn1crypto==1.5.1
# attrs==21.2.0
# bcrypt==3.2.0
# blinker==1.4
# Brlapi==0.8.3
# certifi==2020.6.20
# cffi==1.16.0
# chardet==4.0.0
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.4
# command-not-found==0.3
# cryptography==42.0.5
# cssselect2==0.7.0
# cupshelpers==1.0
# dbus-python==1.2.18
# defer==1.0.6
# distlib==0.3.7
# distro==1.7.0
# distro-info==1.1+ubuntu0.2
# Django==5.0.7
# django-cors-headers==4.4.0
# django-dotenv==1.4.2
# django-environ==0.11.2
# djangorestframework==3.15.2
# djangorestframework-simplejwt==5.3.1
# docker==5.0.3
# docker-compose==1.29.2
# dockerpty==0.4.1
# docopt==0.6.2
# duplicity==0.8.21
# fasteners==0.14.1
# filelock==3.13.1
# future==0.18.2
# gunicorn==22.0.0
# html5lib==1.1
# httplib2==0.20.2
# idna==3.3
# image==1.5.33
# importlib-metadata==4.6.4
# jeepney==0.7.1
# jsonschema==3.2.0
# keyring==23.5.0
# language-selector==0.1
# launchpadlib==1.10.16
# lazr.restfulclient==0.14.4
# lazr.uri==1.0.6
# lockfile==0.12.2
# louis==3.20.0
# lxml==5.1.0
# macaroonbakery==1.3.1
# monotonic==1.6
# more-itertools==8.10.0
# netifaces==0.11.0
# numpy==1.26.4
# oauthlib==3.2.0
# olefile==0.46
# oscrypto==1.3.0
# packaging==24.0
# pandas==2.2.1
# paramiko==2.9.3
# pexpect==4.8.0
# pillow==10.2.0
# platformdirs==3.11.0
# protobuf==3.12.4
# psycopg2==2.9.9
# psycopg2-binary==2.9.9
# ptyprocess==0.7.0
# pycairo==1.20.1
# pycparser==2.21
# pycups==2.0.1
# pygame==2.5.2
# PyGObject==3.42.1
# pyHanko==0.23.0
# pyhanko-certvalidator==0.26.3
# PyJWT==2.3.0
# pymacaroons==0.13.0
# PyNaCl==1.5.0
# pyparsing==2.4.7
# pypdf==4.1.0
# pypng==0.20220715.0
# PyQRCode==1.2.1
# pyRFC3339==1.1
# pyrsistent==0.18.1
# python-apt==2.4.0+ubuntu3
# python-bidi==0.4.2
# python-dateutil==2.9.0.post0
# python-debian==0.1.43+ubuntu1.1
# python-decouple==3.8
# python-dotenv==1.0.1
# pytz==2022.1
# pyxdg==0.27
# PyYAML==6.0.1
# qrcode==7.4.2
# reportlab==4.0.9
# requests==2.31.0
# ruff==0.3.2
# SecretStorage==3.3.1
# six==1.16.0
# sqlparse==0.4.2
# svglib==1.5.1
# systemd-python==234
# texttable==1.6.4
# tinycss2==1.2.1
# typing_extensions==4.10.0
# tzdata==2024.1
# tzlocal==5.2
# ubuntu-drivers-common==0.0.0
# ubuntu-pro-client==8001
# ufw==0.36.1
# unattended-upgrades==0.1
# uritools==4.0.2
# urllib3==1.26.5
# usb-creator==0.3.7
# virtualenv==20.24.6
# wadllib==1.3.6
# webencodings==0.5.1
# websocket-client==1.2.3
# xdg==5
# xhtml2pdf==0.2.15
# xkit==0.0.0
# XlsxWriter==3.2.0
# zipp==1.0.0
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
